Remove "routes" file in api-pls server
